+++ src/main/java/capstondesign2/backend/showStatistics/repository/StatisticsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StatisticsRepository extends JpaRepository<StatisticsEntity, Double> {
+public interface StatisticsRepository extends JpaRepository<StatisticsEntity, Long> {
 }
